restructured files and types
