+++ pages/_app.tsx
@@ -2,7 +2,7 @@ import type { AppProps } from "next/app";
 import dynamic from "next/dynamic";
 import "styles/globals.css";
 const Navbar = dynamic(() => import("src/nav/Navbar"), { ssr: false });
-import { AuthProvider } from "api/AuthProvider";
+import { AuthProvider } from "src/auth/components/AuthProvider";
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
