--- pages/word/index.tsx
@@ -1,12 +1,0 @@
-import type { NextPage } from "next";
-import React, { useEffect, useRef } from "react";
-
-const Word: NextPage = () => {
-  return (
-    <>
-      <div>test</div>
-    </>
-  );
-};
-
-export default Word;
